[HumanBenchmark][Backend][Workouts] refactor workouts endpoints
